fix(ci): resolve test assertion and mypy type errors
- Fix test_returns_numbered_text: match [REF:1] format instead of [1]
- Fix mypy errors: explicit casts for no-any-return in device.py,
  ollama_manager.py, analysis_cache.py, stress_detector.py, expectation_inference.py
- Fix Row.get() attr-defined in repository.py (cast to dict)
- Fix bool not callable in register.py (rename shadowed variable)
- Fix ExpectationViolation call-arg: add character_id/character_name fields
- Fix dict-item type mismatch in stress_detector.py

+++ src/narrative_assistant/core/device.py
@@ -516,7 +516,7 @@ def _get_system_ram_gb() -> float:
     try:
         import psutil
 
-        return psutil.virtual_memory().total / (1024**3)
+        return float(psutil.virtual_memory().total / (1024**3))
     except ImportError:
         pass
